@@ -2,15 +2,6 @@
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
-[tool.black]
-line-length = 100
-target-version = ["py36", "py37", "py38"]
-include = "\\.pyi?$"
-
-[tool.isort]
-profile = "black"
-line_length = 100
-
 [tool.poetry]
 name = "spiel"
 version = "0.3.0"
@@ -24,7 +15,6 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Operating System :: MacOS",
     "Operating System :: Unix",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Topic :: Multimedia :: Graphics :: Presentation",
     "Typing :: Typed"]
@@ -33,9 +23,9 @@ license = "MIT"
 include = ["py.typed", "demo/*"]
 
 [tool.poetry.dependencies]
-python = ">=3.9,<4"
+python = ">=3.10,<4"
 rich = ">=12"
-typer = ">=0.3"
+typer = ">=0.6"
 watchdog = ">=2"
 pendulum = ">=2"
 Pillow = ">=8"
@@ -49,25 +39,34 @@ pytest = ">=6"
 pytest-watch = ">=4"
 pytest-cov = ">=3"
 pytest-xdist = ">=2"
-mypy = ">=0.900"
+mypy = ">=0.961"
 pytest-mypy = ">=0.8"
 pytest-mock = ">=3"
 hypothesis = ">=6"
-Sphinx = ">=4"
-sphinx-rtd-theme = ">=0.5"
-sphinx-autobuild = ">=2021.3.14"
-line-profiler = ">=3"
 
 [tool.poetry.scripts]
 spiel = 'spiel.main:app'
 
+[tool.black]
+line-length = 100
+target-version = ["py39", "py310"]
+include = "\\.pyi?$"
+
+[tool.isort]
+profile = "black"
+line_length = 100
+
+[tool.pycln]
+all = true
+
 [tool.pytest.ini_options]
-addopts = ["--strict-markers", "--mypy"]
+addopts = ["--strict-markers", "--mypy", "-n", "auto"]
 testpaths = ["tests", "spiel"]
 norecursedirs = ["demo"]
 
 [tool.mypy]
 pretty = false
+show_error_codes = true
 
 files = ["spiel/*.py", "tests/*.py"]
 
