@@ -25,10 +25,10 @@ def test_reloader_triggers_when_file_modified(
         file_with_empty_deck.write_text(
             dedent(
                 f"""\
-    from spiel import Deck
+                from spiel import Deck
 
-    {DECK} = Deck(name="modified")
-    """
+                {DECK} = Deck(name="modified")
+                """
             )
         )
 
